feat: Add TransparentChatWindow.toggle_commands method
This commit adds the `TransparentChatWindow.toggle_commands` method to the `gui.py` module. This method is responsible for toggling the display of available commands in the chat window. It is triggered by the voice command "Kommandos" or "1. Kommandos".

The code changes include:
- Adding the `toggle_commands` method to the `TransparentChatWindow` class
- Updating the list of commands and their descriptions to include the "Kommandos" option

This feature enhances the user experience by providing a convenient way to view and access the available commands in the chat window.